component for character card is now available
